@@ -12,7 +12,7 @@ class Milestone < ApplicationRecord
     return if payload.blank?
 
     repo.milestones.find_or_initialize_by(remote_id: payload.fetch(:id)).tap do |milestone|
-      milestone.update_attributes(
+      milestone.update(
         number: payload[:number],
         title: payload[:title],
         state: payload[:state],
