@@ -69,7 +69,7 @@ RSpec.describe TicketMovesController, type: :request do
     end
 
     before do
-      dev.update_attributes!(display_duration: true)
+      dev.update!(display_duration: true)
 
       2.times do
         ticket = create(:ticket, repo: repo)
@@ -110,7 +110,7 @@ RSpec.describe TicketMovesController, type: :request do
             subject
           }.to change { board_ticket.reload.swimlane }.from(backlog).to(dev)
             .and change { dev.reload.board_tickets[destination_position] }.to(board_ticket)
-            .and have_broadcasted_to([:board, board]).from_channel(BoardChannel).with(expected_payload)
+            .and have_broadcasted_to(board).from_channel(BoardChannel).with(expected_payload)
 
           expect(response).to have_http_status(:created)
           expect(JSON.parse(response.body)).to include(
@@ -153,7 +153,7 @@ RSpec.describe TicketMovesController, type: :request do
                 subject
               }.to change { board_ticket.reload.swimlane }.from(backlog).to(dev)
                 .and change { dev.reload.board_tickets[destination_position] }.to(board_ticket)
-                .and have_broadcasted_to([:board, board]).from_channel(BoardChannel).with(expected_payload)
+                .and have_broadcasted_to(board).from_channel(BoardChannel).with(expected_payload)
 
               expect(response).to have_http_status(:created)
 
@@ -180,7 +180,7 @@ RSpec.describe TicketMovesController, type: :request do
                 subject
               }.to change { board_ticket.reload.swimlane }.from(backlog).to(dev)
                 .and change { dev.reload.board_tickets[destination_position] }.to(board_ticket)
-                .and have_broadcasted_to([:board, board]).from_channel(BoardChannel).with(expected_payload)
+                .and have_broadcasted_to(board).from_channel(BoardChannel).with(expected_payload)
 
               expect(response).to have_http_status(:created)
 
