@@ -54,7 +54,7 @@ class ApplicationRecord < ActiveRecord::Base
 
   # We need to capture changes here, since they get wiped
   # if the model is saved again - even if there are no updates
-  # to persist (this can happen if `update_attributes` is called
+  # to persist (this can happen if `update` is called
   # multiple times within the transaction
   def capture_changes_for_broadcast
     @changes_for_broadcast = saved_changes.transform_values(&:first)
