@@ -1,1 +1,1 @@
-import Rails from "rails-ujs";
+import Rails from "@rails/ujs";
